requirements.txt
# Bot dependencies
aiogram==3.4.1
aiosqlite==0.19.0
SQLAlchemy[asyncio]==2.0.25
greenlet==3.2.4

# Web panel dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
